New port layer timeout expiry macros
Two new macros are introduced into u_port.h: U_PORT_TICK_TIME_EXPIRED_TIMEOUT_OR_WRAP_MS() and U_PORT_TICK_TIME_BEYOND_STOP_OR_WRAP_MS().

These macros perform timeout comparisons with signed 32 bit arithmetic to ensure that, should the underlying 32-bit tick counter wrap (which will happen after 25 days), it will be as if the timeout had expired, rather than being as if the timeout won't expire for another 25 days. U_PORT_TICK_TIME_EXPIRED_TIMEOUT_OR_WRAP_MS() is for the case where you have a start tick time in milliseconds and a timeout in milliseconds, U_PORT_TICK_TIME_BEYOND_STOP_OR_WRAP_MS() is for the case where you have just a stop time in milliseconds.

The timeout comparisons throughout the code are changed to use these macros and, on the way, unneeded casts to int32_t, left over from the days when uPortTickTimeMs() returned an int64_t, are also removed.

+++ common/at_client/test/u_at_client_test.c
@@ -309,7 +309,7 @@ static bool atTimeoutIsObeyed(uAtClientHandle_t atClientHandle,
     uPortTaskBlock(10);
     if ((x < 0) && (y < 0) &&
         (gConsecutiveTimeout == consecutiveTimeouts + 1)) {
-        duration = (int32_t) (uPortGetTickTimeMs() - startTimeMs);
+        duration = uPortGetTickTimeMs() - startTimeMs;
         if ((duration < timeoutMs) ||
             (duration > timeoutMs + U_AT_CLIENT_TEST_AT_TIMEOUT_TOLERANCE_MS)) {
             U_TEST_PRINT_LINE("AT timeout was not obeyed (%d ms as opposed"
